+++ src/Gnutella.cc
@@ -12,6 +12,7 @@
 #include "Event.h"
 #include "PIA.h"
 
+#include <algorithm>
 
 GnutellaMsgState::GnutellaMsgState()
 	{
@@ -238,7 +239,7 @@ void Gnutella_Analyzer::SendEvents(GnutellaMsgState* p, bool is_orig)
 		vl->append(new StringVal(p->payload));
 		vl->append(new Val(p->payload_len, TYPE_COUNT));
 		vl->append(new Val((p->payload_len <
-				    min(p->msg_len, GNUTELLA_MAX_PAYLOAD)),
+				    min(p->msg_len, (unsigned int)GNUTELLA_MAX_PAYLOAD)),
 				   TYPE_BOOL));
 		vl->append(new Val((p->payload_left == 0), TYPE_BOOL));
